requirements.txt
Flask==2.3.2
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-caching==2.0.2
gunicorn==20.1.0
pi-network-sdk==0.4.2
